adding index page to render the current data
